Show why free insurance
Extract FreeInsuranceCase component.

Create WhyFreeInsurance component.
