+++ src/locales.json
@@ -27,6 +27,8 @@
     "free_insurance_case": "We found and started a way to provide free insurance for low-income people in India.\r",
     "free_insurance_story": "Jafar Miyan is a 42-year-old father with four dependents.\nHis daily income was around 3,100 Rs. (4 EUR), so he could not afford insurance.\r\nWe gave him free health insurance for 449 Rs. (5.7 EUR) per year.\r\nIf he goes to the hospital because he is not feeling well, he can be guaranteed up to 1,000 Rs. (12.7 EUR) per day for up to 30 days a year.",
     "free_insurance_title": "We're\ndonating\nfree insurance",
+    "free_insurance_why_description": "More than 3.7 billion people are not covered by insurance because they do not have a monthly premium of 0.8 EUR.\nLow-income households have a lower quality of life in the event of a family accident.",
+    "free_insurance_why_title": "Why",
     "heading": "Secure more people equally",
     "key_message1": "*Free* insurance\nfor *low-income*\npeople",
     "key_message2": "Monthly \n*subscription*!\nPay only for the \n*months* you use",
@@ -70,6 +72,8 @@
     "free_insurance_case": "We found and started a way to provide free insurance for low-income people in India.\r",
     "free_insurance_story": "Jafar Miyan is a 42-year-old father with four dependents.\nHis daily income was around 3,100 Rs. (4 EUR), so he could not afford insurance.\r\nWe gave him free health insurance for 449 Rs. (5.7 EUR) per year.\r\nIf he goes to the hospital because he is not feeling well, he can be guaranteed up to 1,000 Rs. (12.7 EUR) per day for up to 30 days a year.",
     "free_insurance_title": "We're\ndonating\nfree insurance",
+    "free_insurance_why_description": "More than 3.7 billion people are not covered by insurance because they do not have a monthly premium of 0.8 EUR.\nLow-income households have a lower quality of life in the event of a family accident.",
+    "free_insurance_why_title": "Why",
     "heading": "Secure more people equally",
     "key_message1": "*Free* insurance\nfor *low-income*\npeople",
     "key_message2": "Monthly \n*subscription*!\nPay only for the \n*months* you use",
